requirements.txt
fastapi
uvicorn
pandas
lightgbm
scikit-learn
requestsd]